pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-bom"
version = "0.11.1"
description = "AI Bill of Materials (AI-BOM) generator — CVE scanning, blast radius, enterprise remediation plans, OWASP LLM Top 10 + MITRE ATLAS threat mapping, and MCP registry for AI agents."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "agent-bom", email = "security@agent-bom.dev"}
]
keywords = ["ai-bom", "sbom", "mcp", "mcp-server", "security", "ai-agents", "vulnerability", "supply-chain", "owasp", "mitre-atlas", "grype", "syft", "blast-radius", "cve", "llm-security", "remediation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "httpx>=0.28.1",  # Requires urllib3>=2.6.3 (CVE-2025-66418, CVE-2025-66471, CVE-2026-21441)
    "pydantic>=2.0",
    "cyclonedx-python-lib>=11.6",  # Updated for better CycloneDX 1.6 support
    "packageurl-python>=0.17",
    "toml>=0.10",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sse-starlette>=2.1",
]
otel = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp-proto-http>=1.20",
]
ui = [
    "streamlit>=1.35",
    "plotly>=5.20",
    "pandas>=2.0",
]
aws = ["boto3>=1.34"]
azure = [
    "azure-identity>=1.16",
    "azure-mgmt-resource>=23.0",
    "azure-ai-projects>=1.0.0b3",
    "azure-mgmt-appcontainers>=3.0",
]
gcp = [
    "google-cloud-aiplatform>=1.48",
    "google-auth>=2.27",
]
databricks = ["databricks-sdk>=0.20"]
snowflake = ["snowflake-connector-python>=3.6"]
nebius = ["nebius>=0.1"]
cloud = [
    "agent-bom[aws]",
    "agent-bom[azure]",
    "agent-bom[gcp]",
    "agent-bom[databricks]",
    "agent-bom[snowflake]",
    "agent-bom[nebius]",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.4",
    "mypy>=1.0",
    "pip-audit>=2.10",  # Security dependency scanning
    "bandit>=1.9",      # Static security analysis
    "safety>=3.7",      # Vulnerability checking
]

[project.scripts]
agent-bom = "agent_bom.cli:main"

[project.urls]
Homepage = "https://github.com/agent-bom/agent-bom"
Repository = "https://github.com/agent-bom/agent-bom"
Issues = "https://github.com/agent-bom/agent-bom/issues"
Changelog = "https://github.com/agent-bom/agent-bom/releases"
Documentation = "https://github.com/agent-bom/agent-bom#readme"
"Security Policy" = "https://github.com/agent-bom/agent-bom/blob/main/SECURITY.md"
"Trust & Permissions" = "https://github.com/agent-bom/agent-bom/blob/main/PERMISSIONS.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 140  # Relaxed from 120 for CLI help strings
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = [
    "E501",  # Line too long (handled by line-length setting)
    "F841",  # Unused variable (will clean up later)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
